Extract used prop types detection code
