+++ lib/util/Components.js
@@ -12,6 +12,7 @@ const pragmaUtil = require('./pragma');
 const astUtil = require('./ast');
 const propTypes = require('./propTypes');
 const jsxUtil = require('./jsx');
+const usedPropTypesUtil = require('./usedPropTypes');
 
 function getId(node) {
   return node && node.range.join(':');
@@ -37,6 +38,7 @@ function mergeUsedPropTypes(propsList, newPropsList) {
       propsToAdd.push(newProp);
     }
   });
+
   return propsList.concat(propsToAdd);
 }
 
@@ -142,7 +144,8 @@ class Components {
         const newUsedProps = (this._list[i].usedPropTypes || []).filter(propType => !propType.node || propType.node.kind !== 'init');
 
         const componentId = getId(component.node);
-        usedPropTypes[componentId] = (usedPropTypes[componentId] || []).concat(newUsedProps);
+
+        usedPropTypes[componentId] = mergeUsedPropTypes(usedPropTypes[componentId] || [], newUsedProps);
       }
     }
 
@@ -154,7 +157,7 @@ class Components {
       const id = getId(this._list[j].node);
       list[j] = this._list[j];
       if (usedPropTypes[id]) {
-        list[j].usedPropTypes = (list[j].usedPropTypes || []).concat(usedPropTypes[id]);
+        list[j].usedPropTypes = mergeUsedPropTypes(list[j].usedPropTypes || [], usedPropTypes[id]);
       }
     }
     return list;
@@ -720,7 +723,9 @@ function componentRule(rule, context) {
   const ruleInstructions = rule(context, components, utils);
   const updatedRuleInstructions = util._extend({}, ruleInstructions);
   const propTypesInstructions = propTypes(context, components, utils);
-  const allKeys = new Set(Object.keys(detectionInstructions).concat(Object.keys(propTypesInstructions)));
+  const usedPropTypesInstructions = usedPropTypesUtil(context, components, utils);
+  const allKeys = new Set(Object.keys(detectionInstructions).concat(Object.keys(propTypesInstructions))
+    .concat(Object.keys(usedPropTypesInstructions)));
   allKeys.forEach(instruction => {
     updatedRuleInstructions[instruction] = function(node) {
       if (instruction in detectionInstructions) {
@@ -728,6 +733,9 @@ function componentRule(rule, context) {
       }
       if (instruction in propTypesInstructions) {
         propTypesInstructions[instruction](node);
+      }
+      if (instruction in usedPropTypesInstructions) {
+        usedPropTypesInstructions[instruction](node);
       }
       return ruleInstructions[instruction] ? ruleInstructions[instruction](node) : void 0;
     };
